+++ src/templates/formtools/wizard/basis_erfassung.html
@@ -1,0 +1,8 @@
+{% extends "formtools/wizard/wizard_form.htm" %}
+
+
+
+{% block feedback_veranstalter_form %}
+    Wir sind hier im Schritt1. Langer text....
+{{ wizard.form }}
+{% endblock %}
