+++ src/templates/formtools/wizard/basis_erfassung2.html
@@ -1,0 +1,8 @@
+{% extends "formtools/wizard/wizard_form.htm" %}
+
+
+
+{% block feedback_veranstalter_form %}
+    Im Schritt 2
+{{ wizard.form }}
+{% endblock %}
