+++ src/templates/formtools/wizard/wizard_form.htm
@@ -1,0 +1,24 @@
+{% extends "bestellung_base.html" %}
+
+{% block head %}
+{{ wizard.form.media }}
+{% endblock %}
+
+{% block content %}
+<p>Step {{ wizard.steps.step1 }} of {{ wizard.steps.count }}</p>
+<form action="" method="post">{% csrf_token %}
+<table>
+{{ wizard.management_form }}
+
+{% block feedback_veranstalter_form %}
+{% endblock %}
+
+</table>
+{% if wizard.steps.prev %}
+<button name="wizard_goto_step" type="submit" value="{{ wizard.steps.first }}">Erster Schritt</button>
+<button name="wizard_goto_step" type="submit" value="{{ wizard.steps.prev }}">"Vorheriger Schritt"</button>
+{% endif %}
+<input type="submit" value="Speichern"/>
+</form>
+{% endblock %}
+
